fix: make uicc suspend an optional build flag
The implementations is no where complete.
However, what is provided might be OK for some cases.

It is mainly interesting on modems that won't go to deep sleep without suspending the SIM first.
It doesn't really make sense in a SoftSIM context all the time - and this is the current compromise.