chore: remove old build system
With the introduction of CMake, we have slowly been
moving towards this ways of configuring the project.

We are removing the previously used build system to
remove the need for dual build system maintenance.

The motivation for this shift is to be a better match
for the many CMake projects in which the onomondo-uicc
can and should be integration as part of.