feat: add option to link against custom crypto implementations
We seek the option to link against custom crypto implementations
to be used for key storage and crypto computational tasks.
This is introduced as an configurable option.

In the work, common crypto part of the codespace
has been moved public include file folder.

Add feature that ensured that the externally loaded keys
are memzero'ed after their use.

Add automated build processes with multiple different
compiler types and CMake flag combinations.
More might be added in the future.

Update the repository documentation for external key loading.

---------

Co-authored-by: Benjamin Bruun <bb@onomondo.com>